feat: Days, Categories enum 생성
