+++ src/main/java/menu/Application.java
@@ -1,7 +1,10 @@
 package menu;
 
+import menu.controller.Controller;
+
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+        Controller controller = new Controller();
+        controller.init();
     }
 }
